# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Industry Monitor API (勐腊数据采集)",
    "description": "勐腊行业数据监控与采集系统 API。支持行业数据查询、定时采集、缓存管理、告警监控等功能。",
    "version": "1.0.0",
    "contact": {
      "name": "Admin"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "本地开发环境"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer Token 认证。通过 /api/auth/login 获取 token。"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": { "type": "string", "description": "用户名", "example": "admin" },
          "password": { "type": "string", "description": "密码", "example": "password123" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "description": "JWT Token" },
          "username": { "type": "string", "description": "用户名" }
        }
      },
      "GenerateTokenRequest": {
        "type": "object",
        "properties": {
          "label": { "type": "string", "default": "api", "description": "Token 标签" },
          "expires_hours": { "type": "integer", "default": 8760, "description": "过期时间（小时），默认1年" }
        }
      },
      "GenerateTokenResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string", "description": "生成的长期 Token" },
          "label": { "type": "string", "description": "Token 标签" },
          "expires_hours": { "type": "integer", "description": "过期时间（小时）" }
        }
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "username": { "type": "string", "description": "用户名" },
          "payload": {
            "type": "object",
            "properties": {
              "sub": { "type": "string", "description": "Subject (用户名)" },
              "iat": { "type": "integer", "description": "签发时间 (Unix timestamp)" },
              "exp": { "type": "integer", "description": "过期时间 (Unix timestamp)" }
            }
          }
        }
      },
      "MengLaQueryBody": {
        "type": "object",
        "required": ["action"],
        "properties": {
          "action": {
            "type": "string",
            "enum": ["high", "hot", "chance", "industryViewV2", "industryTrendRange"],
            "description": "查询动作类型"
          },
          "product_id": { "type": "string", "default": "", "description": "产品 ID" },
          "catId": { "type": "string", "default": "", "description": "类目 ID（需存在于 category.json 中）" },
          "dateType": {
            "type": "string",
            "enum": ["day", "month", "quarter", "year"],
            "description": "日期粒度"
          },
          "timest": { "type": "string", "description": "时间周期 key" },
          "starRange": { "type": "string", "description": "趋势起始范围" },
          "endRange": { "type": "string", "description": "趋势结束范围" },
          "extra": { "type": "object", "description": "额外参数" }
        }
      },
      "MengLaQueryParamsBody": {
        "type": "object",
        "properties": {
          "product_id": { "type": "string", "default": "", "description": "产品 ID" },
          "catId": { "type": "string", "default": "", "description": "类目 ID" },
          "dateType": {
            "type": "string",
            "enum": ["day", "month", "quarter", "year"],
            "description": "日期粒度"
          },
          "timest": { "type": "string", "description": "时间周期 key" },
          "starRange": { "type": "string", "description": "趋势起始范围" },
          "endRange": { "type": "string", "description": "趋势结束范围" },
          "extra": { "type": "object", "description": "额外参数" }
        }
      },
      "HighItem": {
        "type": "object",
        "description": "蓝海行业数据项",
        "properties": {
          "catNameCn": { "type": "string", "description": "类目中文名" },
          "catName": { "type": "string", "description": "类目名称" },
          "catTag": { "type": "integer", "description": "类目标签" },
          "catId1": { "type": "string", "description": "一级类目ID" },
          "catId2": { "type": "string", "description": "二级类目ID" },
          "catId3": { "type": "string", "description": "三级类目ID" },
          "skuNum": { "type": "integer", "description": "SKU 数量" },
          "saleSkuNum": { "type": "integer", "description": "在售 SKU 数量" },
          "saleRatio": { "type": "number", "format": "float", "description": "在售比例" },
          "monthSales": { "type": "number", "format": "float", "description": "月销量" },
          "monthSalesRating": { "type": "number", "format": "float", "description": "月销量评分" },
          "monthSalesDynamics": { "type": "number", "format": "float", "description": "月销量变动" },
          "monthGmv": { "type": "number", "format": "float", "description": "月 GMV" },
          "monthGmvRmb": { "type": "number", "format": "float", "description": "月 GMV（人民币）" },
          "monthGmvRating": { "type": "number", "format": "float", "description": "月 GMV 评分" },
          "monthGmvDynamics": { "type": "number", "format": "float", "description": "月 GMV 变动" },
          "brand": { "type": "string", "description": "品牌名" },
          "brandGmv": { "type": "number", "format": "float", "description": "品牌 GMV" },
          "brandGmvRmb": { "type": "number", "format": "float", "description": "品牌 GMV（人民币）" },
          "brandGmvRating": { "type": "number", "format": "float", "description": "品牌 GMV 占比" },
          "topGmv": { "type": "number", "format": "float", "description": "Top GMV" },
          "topGmvRating": { "type": "number", "format": "float", "description": "Top GMV 占比" },
          "topAvgPrice": { "type": "number", "format": "float", "description": "Top 均价" },
          "topAvgPriceRmb": { "type": "number", "format": "float", "description": "Top 均价（人民币）" }
        }
      },
      "HighResponse": {
        "type": "object",
        "properties": {
          "highList": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/HighItem" },
            "description": "蓝海行业列表"
          }
        }
      },
      "HotResponse": {
        "type": "object",
        "properties": {
          "hotList": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/HighItem" },
            "description": "热销行业列表"
          }
        }
      },
      "ChanceResponse": {
        "type": "object",
        "properties": {
          "chanceList": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/HighItem" },
            "description": "潜力行业列表"
          }
        }
      },
      "IndustryRangeBucket": {
        "type": "object",
        "description": "行业区间分布项",
        "properties": {
          "id": { "type": "string", "description": "区间 ID" },
          "title": { "type": "string", "description": "区间标题" },
          "itemCount": { "type": "integer", "description": "商品数量" },
          "sales": { "type": "number", "format": "float", "description": "销量" },
          "gmv": { "type": "number", "format": "float", "description": "GMV" },
          "itemCountRate": { "type": "number", "format": "float", "description": "商品数量占比" },
          "salesRate": { "type": "number", "format": "float", "description": "销量占比" },
          "gmvRate": { "type": "number", "format": "float", "description": "GMV 占比" }
        }
      },
      "IndustryBrandRateItem": {
        "type": "object",
        "description": "行业品牌占比项",
        "properties": {
          "catId": { "type": "string", "description": "类目 ID" },
          "catName": { "type": "string", "description": "类目名称" },
          "catNameCn": { "type": "string", "description": "类目中文名" },
          "brandGmv": { "type": "number", "format": "float", "description": "品牌 GMV" },
          "brandGmvRate": { "type": "number", "format": "float", "description": "品牌 GMV 占比" },
          "brandItemCount": { "type": "integer", "description": "品牌商品数" },
          "brandItemCountRate": { "type": "number", "format": "float", "description": "品牌商品占比" },
          "brandSales": { "type": "number", "format": "float", "description": "品牌销量" },
          "brandSalesRate": { "type": "number", "format": "float", "description": "品牌销量占比" },
          "typeId": { "type": "string", "description": "类型 ID" }
        }
      },
      "IndustryViewResponse": {
        "type": "object",
        "description": "行业概览数据",
        "properties": {
          "industrySalesRangeDtoList": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/IndustryRangeBucket" },
            "description": "销量区间分布"
          },
          "industryGmvRangeDtoList": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/IndustryRangeBucket" },
            "description": "GMV 区间分布"
          },
          "industryPriceRangeDtoList": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/IndustryRangeBucket" },
            "description": "价格区间分布"
          },
          "industryBrandRateDtoList": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/IndustryBrandRateItem" },
            "description": "品牌占比列表"
          }
        }
      },
      "IndustryTrendPoint": {
        "type": "object",
        "description": "行业趋势数据点",
        "properties": {
          "timest": { "type": "string", "description": "时间标记" },
          "salesSkuCount": { "type": "integer", "description": "在售 SKU 数" },
          "salesSkuRatio": { "type": "number", "format": "float", "description": "在售 SKU 比例" },
          "monthSales": { "type": "number", "format": "float", "description": "月销量" },
          "monthSalesRatio": { "type": "number", "format": "float", "description": "月销量环比" },
          "monthGmv": { "type": "number", "format": "float", "description": "月 GMV" },
          "monthGmvRatio": { "type": "number", "format": "float", "description": "月 GMV 环比" },
          "currentDayPrice": { "type": "number", "format": "float", "description": "当日均价" }
        }
      },
      "IndustryTrendResponse": {
        "type": "object",
        "properties": {
          "industryTrendRange": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/IndustryTrendPoint" },
            "description": "趋势数据点列表"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "detail": { "type": "string", "description": "错误详情" }
        }
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["认证"],
        "summary": "用户登录",
        "description": "使用用户名密码登录，获取 JWT Token。有登录频率限制（60秒内最多10次）。",
        "operationId": "login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": {
            "description": "认证失败",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/generate-token": {
      "post": {
        "tags": ["认证"],
        "summary": "生成长期 Token",
        "description": "为已认证用户生成长期有效的 API Token（默认1年）",
        "operationId": "generateToken",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GenerateTokenRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GenerateTokenResponse" }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["认证"],
        "summary": "获取当前用户信息",
        "description": "验证当前 Token 并返回用户信息",
        "operationId": "getMe",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MeResponse" }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/data/mengla/query": {
      "post": {
        "tags": ["MengLa 数据"],
        "summary": "统一查询接口",
        "description": "勐腊统一查询端点，支持所有 action 类型（high/hot/chance/industryViewV2/industryTrendRange）",
        "operationId": "menglaQuery",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MengLaQueryBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功（响应结构取决于 action 类型）",
            "headers": {
              "X-MengLa-Source": {
                "description": "数据来源 (l1/l2/l3/fresh)",
                "schema": { "type": "string", "enum": ["l1", "l2", "l3", "fresh"] }
              },
              "X-MengLa-Trend-Partial": {
                "description": "趋势数据部分返回标记 (requested,found)",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/data/mengla/high": {
      "post": {
        "tags": ["MengLa 数据"],
        "summary": "蓝海行业 Top",
        "description": "查询蓝海行业 Top 数据",
        "operationId": "menglaHigh",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MengLaQueryParamsBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-MengLa-Source": {
                "description": "数据来源",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HighResponse" }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/data/mengla/hot": {
      "post": {
        "tags": ["MengLa 数据"],
        "summary": "热销行业 Top",
        "description": "查询热销行业 Top 数据",
        "operationId": "menglaHot",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MengLaQueryParamsBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-MengLa-Source": {
                "description": "数据来源",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HotResponse" }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/data/mengla/chance": {
      "post": {
        "tags": ["MengLa 数据"],
        "summary": "潜力行业 Top",
        "description": "查询潜力行业 Top 数据",
        "operationId": "menglaChance",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MengLaQueryParamsBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-MengLa-Source": {
                "description": "数据来源",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChanceResponse" }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/data/mengla/industry-view": {
      "post": {
        "tags": ["MengLa 数据"],
        "summary": "行业概览",
        "description": "查询行业概览数据（销量/GMV/价格区间分布 + 品牌占比）",
        "operationId": "menglaIndustryView",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MengLaQueryParamsBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-MengLa-Source": {
                "description": "数据来源",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IndustryViewResponse" }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/data/mengla/industry-trend": {
      "post": {
        "tags": ["MengLa 数据"],
        "summary": "行业趋势",
        "description": "查询行业趋势数据（通常使用 starRange/endRange 指定时间范围）",
        "operationId": "menglaIndustryTrend",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MengLaQueryParamsBody" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "headers": {
              "X-MengLa-Source": {
                "description": "数据来源",
                "schema": { "type": "string" }
              },
              "X-MengLa-Trend-Partial": {
                "description": "趋势数据部分返回标记",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/IndustryTrendResponse" }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}